Frogger_AI/requirements.txt
cycler==0.10.0
graphviz==0.16
kiwisolver==1.3.1
matplotlib==3.3.4
neat-python==0.92
numpy==1.20.1
Pillow==8.1.0
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
